Autotools: fix: use '$(OBJEXT)' abstraction instead of '.o'
The `OBJEXT` variable contains the extension for object files on the target
platform. The Autotools docs do not seem to discuss this directly, but
third-party sources¹ note:

  `AC_OBJEXT`
    Sets the output variable `OBJEXT` to the extension of object files produced
    by the compiler. Usually, it is set to `.o' on Unix systems and `.obj' on
    Windows.

The Graphviz Autotools files do not call `AC_OBJEXT`, but apparently this is
invoked by `AC_PROG_CC`.²

The effect of this problem was probably that dependency information was missing
on e.g. Cygwin, causing parallel compilations to spuriously fail.

¹ https://www.star.bnl.gov/~liuzx/autobook.html
² https://sourceware.org/legacy-ml/autoconf/2000-12/msg00127.html