fix pkgIndex.tcl being host-system-based
pkgIndex.tcl is constructed by the shell script tclpkg/mkpkgindex.sh that was
using `uname` to decide what the suffix of shared libraries is. `uname` looks at
the current host system. During cross-compilation, this does not necessarily
correspond to the target system where Graphviz will eventually be deployed. So,
e.g. during a Linux compilation targeting macOS, the TCL binding’s shared
libraries would incorrectly be thought to end in “.so” though they actually end
in “.dylib”.

This change fixes the problem by moving this logic into the build systems that
have a more nuanced understanding of the multiple environments involved.

3fcc3c62e7d8fda22ca021e2317bd2cb567b8113 compounded the original problem by
making the `sed` expression in this script sometimes operate on binary files,
possibly producing false positives or false negatives. This was observed
recently,¹ but the present change side steps this in removing this logic
altogether.

Surprisingly Autotools does not seem to have a built-in variable for the shared
library suffix. CMake does, and in fact has a more nuanced understanding of how
a shared library may be named, with a `$CMAKE_SHARED_LIBRARY_PREFIX` as well as
`$CMAKE_SHARED_LIBRARY_SUFFIX`. This change ignores prefixes for now, preserving
the previous (possibly incomplete/incorrect) behavior.

Note that it is correct that these changes insert this logic into a
`$host_os`-based switch in the Autotools build system, though it seems
intuitively it should be based on `$target_os`. Autotools has a
counter-intuitive interpretation of these concepts:

  • `$build_os` – where the code is compiled
  • `$host_os` – where the compiled code will run
  • `$target_os` – what environment the compiled code will produce for

The last of these is only relevant when the software you are compiling is itself
a compiler.²

The HP-UX case introduced in the Autotools build system is just a copy-paste of
the `*` fall through case. I have not checked whether the values of `BROWSER`
and `JSHEXT` here are reasonable and I do not have an environment on which to
test. But this preserves the previous behavior HP-UX users would have.

¹ https://gitlab.com/graphviz/graphviz/-/merge_requests/3937#note_2033742597
² Though counter-intuitive, Autotools’ decision here is well-motivated. It is
  trying to support the most general scenario, running a so-called “Canadian
  Cross” build. https://en.wikipedia.org/wiki/Cross_compiler#Canadian_Cross