cgraph: overload 'agxbuf.located' to serve as a terminator for short strings
The `agxbuf` type implements a dynamically-growing string, complete with Small
String Optimization (SSO).¹ Until now it managed to pack, on x86-64, at most 30
bytes (excluding `'\0'` terminator) into the inline storage. This change
increases this to 31.

This may seem significant complexity for a comparatively insignificant
improvement. But the assumption is that many of these buffers are created and
used when processing large graphs.

This SSO implementation heavily prioritizes memory over speed. That is, its
ability to store bytes inline is optimized at the expense of branches within
code to append and use the string. Extending Raymond Chen’s comparison of more
mainstream SSO implementations:²

  ┌──────────────┬─────────┬─────────┬──────────┬─────────┬─────────┐
  │   sizeof /   │   gcc   │   msvc  │   clang  │  agxbuf │  agxbuf │
  │ SSO capacity │         │         │          │  before │  after  │
  ├──────────────┼─────────┼─────────┼──────────┼─────────┼─────────┤
  │ 32-bit mode  │ 24 / 15 │ 24 / 15 │ 12 / 10³ │ 16 / 14 │ 16 / 15 │
  │ 64-bit mode  │ 32 / 15 │ 32 / 15 │ 24 / 22  │ 32 / 30 │ 32 / 31 │
  └──────────────┴─────────┴─────────┴──────────┴─────────┴─────────┘

¹ not a canonical definition, but some explanation at
  https://pvs-studio.com/en/blog/terms/6658/
² https://devblogs.microsoft.com/oldnewthing/20240510-00/?p=109742
³ Raymond’s table says 11 here, but I believe this is wrong. Examining the Clang
  data structure suggests the capacity without the null terminator is 10.