cgraph: avoid heap allocation in 'agxbuf' edge case
`vsnprintf` and friends unavoidably append a `'\0'`, which `agxbprint` was
correctly accounting for. Unfortunately though this led to a counter-intuitive
situation:

  const char stuff[] = "1234567890123456789012345678901";

  agxbuf buf1 = {0};
  agxbput(&buf1, stuff);
  // buf1 still using inline storage; no heap allocation

  agxbuf buf2 = {0};
  agxbprint(&buf2, "%s", stuff);
  // buf2 using heap-allocated storage

The 31-byte (excluding `'\0'` terminator) string here can fit inline but the
`'\0'` from `vsnprintf` was forcing a transition to heap storage.

This change introduces a separate path for this case that uses double buffering
to avoid this issue.