CMake: expand construction of 'gc' target and rename to 'gc_bin'
Upcoming changes add support to the CMake build system for generating the Guile
language bindings. Unfortunately one of the supporting Guile link libraries is
named “libgc”. Adding this to the CMake build system results in something
semantically equivalent to `target_link_libraries(gv_guile PRIVATE gc)` that
CMake interprets as a request to link against the `gc` binary and errors out.
The simplest way of resolving this conundrum seems to be to rename the existing
`gc` target.