tcldot: squash -Wdiscarded-qualifiers warning with 'agopen'
The background of this change is that Tcl 8.4 changed the Tcl API with
respect to constness¹, while allowing users to continue to use the old
API by defining `USE_NON_CONST`. In Tcl 9.0, the use of `USE_NON_CONST`
has been removed and only the new API is available. Adapting to this
change on platforms that use Tcl 8.x, includes the removal of the
definition of `USE_NON_CONST`.

In Tcl 8.x, if `USE_NON_CONST` is _not_ defined, `Tcl_GetStringResult`
returns a `const char *` instead of a `char *`. Strangely, in version
9.0 it again returns `char *`.

To adapt to this in Tcl version 8.x, since "`agopen`: cannot take a
`const char*`"³, we make a temporary non-const copy of the string for
now and look into further modernization later².

The squased warning is:

/home/magjac/graphviz/tclpkg/tcldot/tcldot.c:53:16: warning: passing argument 1 of ‘agopen’ discards ‘const’ qualifier from pointer target type [-Wdiscarded-qualifiers]

GitLab: #2581

¹ https://wiki.tcl-lang.org/page/Migration+to+8.4%3A+CONSTification
² https://gitlab.com/graphviz/graphviz/-/merge_requests/3937#note_2039737586
³ https://gitlab.com/graphviz/graphviz/-/issues/634#note_618052943