tcldot: add 'tcldot_argv_dup' and 'tcldot_argv_free'
In Tcl 9.0 and in earlier versions if `USE_NON_CONST`¹ is _not_ defined,
the `Tcl_CmdProc` `argv` parameter is changed from being declared as
`char *argv[]` to `const char *argv[]`.

As an alternative to letting this change propagate into the internals of
the Graphviz code base, which would require making much more invasive
changes, the new functions will be used in upcoming commits to adapt to
it. This allows us to confine the needed adaptations to `tclpkg` for now
and look into further modernization later¹.

One example of something that would require a much more invasive change
is that "`agopen`: cannot take a `const char*`"².

GitLab: #2581

¹ https://gitlab.com/graphviz/graphviz/-/merge_requests/3937#note_2039737586
² https://gitlab.com/graphviz/graphviz/-/issues/634#note_618052943