tcldot nodecmd: squash -Wincompatible-pointer-types warnings
In Tcl 9.0 and in earlier versions if `USE_NON_CONST`¹ is _not_ defined,
the `Tcl_CmdProc` `argv` parameter is changed from being declared as
`char *argv[]` to `const char *argv[]`.

As an alternative to letting this change propagate into the internals of
the Graphviz code base, which would require making much more invasive
changes, we make a temporary non-const copy of `argv` and use that. This
allows us to confine the needed adaptations to `tclpkg` for now and look
into further modernization later².

GitLab: #2581

¹ https://wiki.tcl-lang.org/page/Migration+to+8.4%3A+CONSTification
² https://gitlab.com/graphviz/graphviz/-/merge_requests/3937#note_2039737586