util: update comment about 'noreturn' portability
The previous intent was to replace the `NORETURN` macro with `_Noreturn` when
moving to C11. However, C11 and C++ standardized different forms for no-return,
`_Noreturn` vs `[[noreturn]]`. So this replacement would not actually decrease
complexity/verbosity here. C23 standardizes the same syntax as C++, so replacing
this with something universal should be possible then.