cgraph generic list: allow allocation base and first element to differ
This change has no immediate effect, but prepares for allowing the list to model
a FIFO queue. It could already model a C++-vector-style list and a stack, but
using it to model a FIFO queue would require an expensive `memmove` on every
pop. This change prepares to support popping from the front being implemented by
incrementing the head location.

The advantage of this yet-to-come queue overloading on this data structure over
lib/cgraph/queue.h is type safety. The generic list’s interfaces are all
expressed in terms of a parametrized type rather than `void *`, allowing safer
usage.

An indirect goal of this change is preserving the performance characteristics of
the generic list’s operations when it is in use as a list or stack. The only
operations that become algorithmically more expensive in this change are those
that now call `*_sync`. And `*_sync` itself is O(1) if the caller has never used
the (not yet implemented) queue functionality.

This change surveys all usage of interior pointers into lists (obtained through
`*_front`, `*_at`, or `*_back`) and inserts `*_sync` calls where appropriate to
preserve correctness. All of these `*_sync` calls are no-ops, and will remain so
even after upcoming changes in this series, except those on `nodelist_t`. All
uses except the `nodelist_t` operations never pop from the front of the list.

Due to the use of >1-byte unicode characters in comments, clang-format chose to
reformat the line end escapes. It is thus difficult to see that this commit is
adding `*_sync` and `*_is_contiguous` functions. It may be beneficial for
reviewers to use `-w` when viewing the diff.