tcldot: temporarily use 'CONST84' when declaring 'argv'
In Tcl version 8.4 and later but less than 9.0, `CONST84` is defined
to nothing if `USE_NON_CONST` is defined and to `const` otherwise¹. In
Tcl version 9.0 it is not defined at all, but in that case it is
defined to nothing in `tcldot.h`.

This change allows us to make stepwise adaptations to Tcl 9.0 and an
upcoming change that removes the use of `USE_NON_CONST`, while not
producing new warnings in any of the intermediate steps before finally
removing the use of `USE_NON_CONST`.

GitLab: #2581

¹ https://wiki.tcl-lang.org/page/Migration+to+8.4%3A+CONSTification