tclpkg: replace 'CONST' with 'const'
In Tcl version >= 8.4 but less than 9.0, `CONST` is defined to `const`
unless `NO_CONST` is defined¹. We do not do that and 8.x is the only
version we currently use, so this is equivalent.

As a side note, while developing support for using Tcl 9.0 on Windows,
I noticed that the code compiled fine even before this change. It
turns out that `CONST` is also defined to `const` by one of the
Windows SDK headers².

¹ There doesn't seem to exist any documentation on `NO_CONST´, but
  https://core.tcl-lang.org/tips/doc/trunk/tip/503.md at least
  mentions its existence.
² C:\Program Files (x86)\Windows Kits\10\Include\10.0.22621.0\shared\minwindef.h