gd plugin set_penstyle: fix pen styles
This replicates a054876cb42d1471a11e9f995940d651355a6cbb which never seems to
have been propagated to the VRML code. This includes the fixup earlier in this
series of reducing the length passed to `gdImageSetStyle` to 12.

Note that a054876cb42d1471a11e9f995940d651355a6cbb, contrary to what its commit
message may suggest, appears to be a bug fix. Observe that it makes the “dashed”
case of the conditional actually dashed again; previously `gdImageSetStyle` was
receiving a 20-entry uniform array of the pen color.