gd plugin gdgen_set_penstyle: remove useless 'dashstyle' writes
The following call to `gdImageSetStyle` passes 12, so all elements beyond this
index are ignored. This appears to have been a typo in
a054876cb42d1471a11e9f995940d651355a6cbb. It is unclear which number is the typo
– whether the author intended to reduce the array writes to 12 or to pass 14 to
`gdImageSetStyle` – so we choose to preserve the current semantics, given that
no users seem to have reported bugs with this.

Note that a054876cb42d1471a11e9f995940d651355a6cbb, contrary to what its commit
message may suggest, appears to be a bug fix. Observe that it makes the “dashed”
case of the conditional actually dashed again; previously `gdImageSetStyle` was
receiving a 20-entry uniform array of the pen color.