gd plugin gvdevice_gd_putBuf: address mismatch with 'gvwrite'
Squashes a -Wsign-conversion warning and a -Wconversion warning.

The `assert(result <= (size_t)len)` assertion here subsumes
`assert(result <= INT_MAX)`. We expect `result` to not only be within the bounds
of an `int` but guaranteed to be at most the number of bytes we requested to
write due to the semantics of `gvwrite`.