gd plugin gd_format: remove pointer alignment assumptions
This code was treating the `imagedata` pointer as if pointing to
`unsigned int`s. While it is likely this pointer came from `malloc` and
therefore is sufficiently aligned, there is no guarantee of this. Reading
through an unaligned pointer is Undefined Behavior.

This also removes an assumption that the platform is little endian when reading
out `imagedata` as `unsigned int`. However this probably has no relevant effect
on portability because GD seems to assume a little endian platform.

The code (both before and after this change) assumes that
`sizeof(unsigned) >= 4`, something that is not universally true. However this
also appears to be assumed by the GD API.

This change also aligns more closely with the GD API, squashing 2
-Wsign-conversion warnings.