smyrna glcreate_gl_topview_menu: remove creation of unused “select edges” button
e23bcdb7158352619139eadae30ed446b033087d seems to have added creation of this
button that is never actually added to the UI (observe how `b` is not added to a
panel before it is reassigned). Given this change was in 2010 and there have
been no reports of this problem, presumably the current layout is what users
want/expect.

This was discovered by ASan reporting the lost reference to the created button
as a memory leak.

Reported-by: Address Sanitizer