smyrna glcreate_gl_topview_menu: remove creation of unused “details” button
This is quite strange, but 019171008aeccdbc1b320a11b5773529d3a2d314 seems to
have added creation of a button that is never actually added to the UI (observe
how `b` is not added to a panel before it is reassigned). Given this change was
in 2009 and there have been no reports of this problem, presumably the current
layout is what users want/expect.

This was discovered by ASan reporting the lost reference to the created button
as a memory leak.

Reported-by: Address Sanitizer