glcomp: introduce base class member to drawable objects
The glcomp drawable objects relied on their first members being a `glObjType`
followed by a `glCompCommon` because pointers to these objects are cast to
`glCompObj` pointers. Not only is this fragile (moving or deleting these members
will not cause compilation failure but will result in runtime crashes), but it
is Undefined Behavior. Casting pointers to these objects to `glCompObj` pointers
was violating strict aliasing.