common: demote 'Y_off' global to a local that is passed around
This is a win in most scenarios, but is a lateral move for xdot rendering. In
this latter scenario, we move `Y_off` from a common global to a core plugin
global. Future changes should investigate moving `xdot_state_t` allocation to a
new `xdot_begin_job` function and then store the resulting pointer in
`job->context` instead of using it as a global.

Gitlab: #2558