common: remove runtime modification of 'fspanDisc'
The destructor for `fspanList` was being toggled to allow clearing the list
while not freeing contained spans that had been shallow-copied into a new list,
`fl`. We can do this in a simpler way by blanking the spans we have copied into
the new list, thus avoiding their freeing when the destructors run.

Note that it is unnecessary to retain the entries in the original list (`ispan`
in this code) because it is cleared immediately following the touched block in
`mkText`.

Gitlab: #2558, !3976