Autotools: remove support for '-perl' to SWIG
It is unclear to me why d362627bead7c2f006985b2e27719a4b82bba5ce structured
detection logic this way. While it seems true that `-perl` and `-perl5` are
aliases in contemporary SWIG, `-perl5` has existed since the very first 1.0 beta
1 release of SWIG on 1996-04-10.¹

Commit df86ec5af6985fc1ada3945237dcf34113db6512² of SWIG even suggests `-perl5`
is to be preferred:

  Consistent use of target language version of options.

  If a version number is supported by the target language prefer to show it in
  the -help. For example, -perl5 instead of -perl.

The clearer structure seems to be to _only_ support `-perl5` that every known
version of SWIG understands.

¹ See the SWIG changelog, https://www.swig.org/Release/CHANGES.
² https://github.com/swig/swig/commit/df86ec5af6985fc1ada3945237dcf34113db6512