common: fix: do not enclose #included render.h headers in 'extern "C"'
When relocating the cgraph/alloc.h header to the bottom of this #include list,
it exposes a latent problem where a C++ source (gvpack.cpp) is #including
render.h which then transitively #includes exit.h that conditionally #includes
C++ headers containing templates:

  In file included from /usr/include/c++/13/bits/stringfwd.h:40,
                   from /usr/include/c++/13/iosfwd:41,
                   from /usr/include/c++/13/ios:40,
                   from /usr/include/c++/13/ostream:40,
                   from /usr/include/c++/13/iostream:41,
                   from /builds/smattr/graphviz/cmd/tools/../../lib/util/exit.h:9,
                   from /builds/smattr/graphviz/cmd/tools/../../lib/cgraph/alloc.h:23,
                   from /builds/smattr/graphviz/cmd/tools/../../lib/cgraph/list.h:12,
                   from /builds/smattr/graphviz/cmd/tools/../../lib/common/globals.h:18,
                   from /builds/smattr/graphviz/cmd/tools/../../lib/common/render.h:36,
                   from /builds/smattr/graphviz/cmd/tools/gvpack.cpp:31:
  /usr/include/c++/13/bits/memoryfwd.h:64:3: error: template with C linkage
     64 |   template<typename>
        |   ^~~~~~~~
  /builds/smattr/graphviz/cmd/tools/../../lib/common/render.h:18:1: note: 'extern "C"' linkage started here
     18 | extern "C" {
        | ^~~~~~~~~~
  …

The problem here is that headers should not put their #includes into a
`extern "C"` scope. This enforces C linkage on declarations that may not have
anticipated nor want this.

The present commit only fixes one instance of many of this scenario in Graphviz.