common: split 'pitem' into two types
The HTML table construction code was using the `pitem` type to store both rows
and cells. This was unnecessary and led to significantly harder to understand
code – when operating on a `pitem` pointer it was not explicitly clear whether
it pointed to a row or cell.

This change improves type safety and readability. Note that we are able to more
precisely type the `Dtdisc_t.size` fields; only a small improvement in clarity,
but still worthwhile.