common: replace 'row_t.rp' cdt queue with generic list
This improves type safety (no more casting of `dtflatten`/`dtlink` results),
heap fragmentation (the `cell_t` wrapper type and its allocation are gone), and
performance (collection elements are contiguous instead of pointer-linked; the
loop to count cells in the current row is removed¹).

In contrast to most other generic lists of pointers, the allocation and
deallocation of the contained pointers is managed externally. These
`htmlcell_t *`s need to outlive the list itself.

Note that no initialization of the list is required (in `addRow`). The zeroing
performed by the preceding `gv_alloc` leaves a valid empty list.

¹ It is somewhat unclear why this loop existed to begin with. cdt has a `dtsize`
  function that seems like it could have been substituted here.