ortho: use a generic list for adjacency list
The vertex adjacency list was implemented as an ordered set. On the surface,
this is correct – the operations we need are iteration and membership checks.
However, an implicit performance assumption was that the number of items in each
set is relatively large. This is not true of the expected outgoing edge count of
each vertex; we expect this to be relatively small. With this in mind, using a
simple list is likely to be faster. We can search small lists faster than
hashing then traversing a tree structure.