common processTbl: use a bitarray for row index set
The row indices are stored as `uint16_t`s. So instead of using a cdt set data
structure, we can use an 8KB bit set to both store data more compactly and
access it faster. A smart compiler should even be able to observe the lifetime
of the backing heap allocation is enclosed and stack-allocate the array on
platforms that can cope with an 8KB stack allocation.

As this is the last use of the integer set code, this commit removes that too.