Move functions to the bottom in htmlparse.y
The HTMLState struct will be moved into htmlparse.h
in the future, because it will be merged with htmllexstate_t
into a common state context for both parser and lexer.

All functions which uses HTMLState have to be defined after
htmlparse.h is included in htmlparse.c,
which is right after the `%{}` block.