tclpkg: remove unnecessary #include of config.h
This seems to have been introduced in 723215da995ab04f87d940b1f937baf1bc8869cc
to work around a Glibc bug. It is unnecessary in a contemporary environment and
causes problems when trying to integrate the PHP bindings into the CMake build
system that is more precise about what is and is not in the include path.