tclpathplan: store index within a vgpane handle
The handling of “vgpanes” (also called “vgcanvas” at times) involved reversing
pointer to its index within a list. This was done by tightly coupling this logic
to the tclhandle implementation; relying on the layout of its internal metadata.
We can avoid this assumption by storing the index the handle allocator gives
when creating a new entry.