Tclpathplan: fix: anticipate a trailing '%' in triangulation callback bindings
The logic for parsing the binding being installed for the triangulation callback
assumed that any '%' was the leader character of an escape sequence. The only
recognized escape sequences are '%r' and '%t', but it was assumed that any other
(invalid) escape sequence should be output with the leader dropped (e.g. '%s' →
's'). Unfortunately this did not account for there being no next character. In
this scenario, the string terminator would be skipped over.

Gitlab: fixes #2596