Tclpathplan: fix segfaults during triangulation of invalid polygons
It is possible to define a <3-point polygon and then request triangulation of
this polygon. This operation bottomed out in pathplan’s `triangulate` which
assumes the input polygon has ≥ 3 points. These assumptions would cause the TCL
interpreter to crash, as out-of-bounds reads were performed on the `pointp`
parameter.

One may reasonably ask why it is possible to define a <3-point polygon or what
even is a <3-point polygon. But it appears deliberate that Tclpathplan allows
users to construct such things. This change chooses to leave the construction of
such malformed polygons enabled, as it is unclear what other effects removing
them may have.