add changelog entry for Tclpathplan bug previously silently fixed
While investigating an orthogonal bug, it was discovered that prior to
38d0aa9d44a0863590821cdd9102dbf50c27a94b during deletion of an existing vgpane
(e.g. `vgpane0 delete`) the `free(tclhandleFree(vgpaneTable, argv[0]))` call
attempts to free an invalid pointer.

The root cause here appears to be an indirection mistake. The originally
allocated pointer is stored in the user data area. But the call to `free` is
passing the address of the user data area. That is, the call to `free` should
have been `free(*(vgpane_t **)tclhandleFree(vgpaneTable, argv[0]))`.