CI: also remove older Graphviz and its dependents on Rocky
Graphviz 2.40.1 was being installed on Rocky 8 and Graphviz 2.44.0 was being
installed on Rocky 9. The checks that used to be in ci/install-packages.sh and
were relocated in the previous commit should have been adapted to also apply to
Rocky in 5f907e9e8a1a4df7bf1205ebfee0becce2c3ca16 and
8493856a583871835305d9579bcf886ed4cc79a1. Their absence potentially masked bugs
with Rocky packages in the past and is impeding work on #2596.

On Rocky 8, this removes 35MB:
  • graphviz 2.40.1-45.el8, 7.9MB
  • asciidoc 8.6.10-0.5.20180627gitf7c2274.el8, 790KB
  • docbook-dtds 1.0-69.el8, 8.3MB
  • docbook-style-xsl 1.79.2-9.el8, 16MB
  • libXaw 1.0.13-10.el8, 494KB
  • libxslt 1.1.32-6.el8, 734KB
  • sgml-common 0.6.3-50.el8, 168KB
  • xorg-x11-fonts-ISO8859-1-100dpi 7.5-19.el8, 1.0MB

On Rocky 9, this removes 40MB:
  • graphviz 2.44.0-26.el9, 12MB
  • asciidoc 9.1.0-3.el9, 949KB
  • docbook-dtds 1.0-79.el9, 8.3MB
  • docbook-style-xsl 1.79.2-16.el9, 16MB
  • libXaw 1.0.13-19.el9, 498KB
  • libXmu 1.1.3-8.el9, 184KB
  • libxslt 1.1.34-9.el9, 754KB
  • sgml-common 0.6.3-58.el9, 168KB
  • xorg-x11-fonts-ISO8859-1-100dpi 7.5-33.el9, 1.0MB