quartz plugin quartzgen_begin_page: do not share memory mappings
Writes to the temporary file being used as swap space here do not need to be
propagated to other processes. Using a private mapping is cheaper, and even
allows a clever implementation to only perform these writes in-memory.

As an aside, it is unclear whether these code paths are still needed. They are
for supporting extremely old versions of iOS that had only limited virtual
memory support. Every iPhone/iPad released in the last 10 years supports more
advanced heap allocation. Moreover all of these devices offer a
`CGBitmapContentCreate` API that allows instead passing in `NULL` and letting
the framework manage this memory instead. Furthermore, it is unclear to me there
are any contemporary users attempting to compile Graphviz for these platforms.
Nevertheless this code is left intact.