CI: suppress Boost warnings about its own deprecated usages
Compilation of the Catch2-based test suite triggers some warnings from Boost
code:

  [ 84%] Building CXX object tests/CMakeFiles/test_common.dir/svgpp_context.cpp.o
  In file included from /usr/include/boost/polymorphic_cast.hpp:58,
                   from /usr/include/boost/range/detail/any_iterator_wrapper.hpp:13,
                   from /usr/include/boost/range/detail/any_iterator.hpp:22,
                   from /usr/include/boost/range/any_range.hpp:17,
                   from /builds/graphviz/graphviz/tests/svgpp_context.h:5,
                   from /builds/graphviz/graphviz/tests/svgpp_context.cpp:8:
  /usr/include/boost/spirit/include/phoenix.hpp:12:1: note: '#pragma message: This header is deprecated. Use <boost/phoenix.hpp> instead.'
     12 | BOOST_HEADER_DEPRECATED("<boost/phoenix.hpp>")
        | ^~~~~~~~~~~~~~~~~~~~~~~

It is unclear what one is supposed to do about this. Boost is complaining about
use of a deprecated API, but the call itself is originating from Boost’s own
code. The macro added in this commit is reportedly the way to suppress these.