Introduce aagscan_t and aagextra_t.
aagscan_t will be an opaque struct generated by flex,
aagextra_t will contain the whole state of the parser
and lexer in the future.
The goal is to make the scanner reentrant,
so that it does not depend on global state.
All static variables in scan.l and grammar.y will be moved
into this struct.