Enable "reentrant" option for scan.l and grammar.y
Flex and bison are both asked to generate reentrant parser and lexer.
Parts of the lexer state is moved into aagscan_t, which is an opaque struct.
aagextra_t will be used in the future to contain the rest of the state,
which is now managed by many static variables.