glcomp glCompButtonNew: fix memory leak of fonts
In CI with leak detection enabled, ASAn reports numerous errors of the form:

  Indirect leak of 13 byte(s) in 1 object(s) allocated from:
      #0 0x7a2dd3bbf748 in strdup ../../../../src/libsanitizer/asan/asan_interceptors.cpp:578
      #1 0x5bcdaee53771 in gv_strdup /builds/smattr/graphviz/lib/glcomp/../util/alloc.h:103
      #2 0x5bcdaee54619 in glNewFont /builds/smattr/graphviz/lib/glcomp/glcompfont.c:61
      #3 0x5bcdaee567e4 in glNewFontFromParent /builds/smattr/graphviz/lib/glcomp/glcompfont.c:108
      #4 0x5bcdaee64156 in glCompInitCommon /builds/smattr/graphviz/lib/glcomp/glcompset.c:177
      #5 0x5bcdaee64387 in glCompSetNew /builds/smattr/graphviz/lib/glcomp/glcompset.c:188
      #6 0x5bcdaede9644 in glcreate_gl_topview_menu /builds/smattr/graphviz/cmd/smyrna/gui/glcompui.c:218
      #7 0x5bcdaee44d5d in init_viewport /builds/smattr/graphviz/cmd/smyrna/viewport.c:320
      #8 0x5bcdaee0adad in main /builds/smattr/graphviz/cmd/smyrna/main.c:378
      #9 0x7a2dd0aa61c9  (/lib/x86_64-linux-gnu/libc.so.6+0x2a1c9) (BuildId: 6d64b17fbac799e68da7ebd9985ddf9b5cb375e6)
      #10 0x7a2dd0aa628a in __libc_start_main (/lib/x86_64-linux-gnu/libc.so.6+0x2a28a) (BuildId: 6d64b17fbac799e68da7ebd9985ddf9b5cb375e6)
      #11 0x5bcdaedab994 in _start (/usr/bin/smyrna+0x2eb994) (BuildId: bc9e8b9413236b2918cc29182131c4994a9a9ce7)

The problem seems to stem from the earlier call to `glCompInitCommon` in this
function already having created a font. The direct call to `glNewFontFromParent`
then overwrites the pointer to it, losing this memory. We could perhaps try to
avoid allocating the earlier font, but it seems simpler and more robust to just
clear any existing font before this initialization. Note that `glDeleteFont` on
a zeroed `glCompFont` is a no-op.

Following this there are still other leaks of this `fontdesc` allocation. But
these will need to be dealt with in a separate future series.

Reported-by: Address Sanitizer