rsvg plugin gvloadimage_rsvg…: use 'rsvg_handle_render_document' where available
Compilation under librsvg ≥ 2.46 was producing warnings:

    CC       gvloadimage_rsvg.lo
  gvloadimage_rsvg.c: In function ‘gvloadimage_rsvg_cairo’:
  gvloadimage_rsvg.c:108:9: warning: ‘rsvg_handle_render_cairo’ is deprecated:
    Use 'rsvg_handle_render_document' instead [-Wdeprecated-declarations]
    108 |         rsvg_handle_render_cairo(rsvgh, cr);
        |         ^~~~~~~~~~~~~~~~~~~~~~~~
  In file included from /usr/include/librsvg-2.0/librsvg/rsvg.h:1464,
                   from gvloadimage_rsvg.c:19:
  /usr/include/librsvg-2.0/librsvg/rsvg-cairo.h:95:10: note: declared here
     95 | gboolean rsvg_handle_render_cairo (RsvgHandle *handle, cairo_t *cr);
        |          ^~~~~~~~~~~~~~~~~~~~~~~~

This change detects when the newer replacement API is available and uses it
instead, removing this warning.

This change was constructed by consulting librsvg’s forward porting advice.¹
While the advice is extensive, it does not precisely tell you how to migrate an
existing `rsvg_handle_render_cairo` call to `rsvg_handle_render_document`. In
particular, the code following this change uses Cairo APIs to scale the SVG to
the desired size and then the librsvg API to constrain it to an identically
sized viewport. Specifying the desired output dimensions twice like this seems
odd, but empirically produces the correct result. It is unclear to me if either
of these steps can be elided. IMHO it would make much more sense to allow the
`viewport` argument to `rsvg_handle_render_document` to be null, indicating a
(0,0)-originating viewport whose dimensions should be derived from the
pre-scaled image, but this does not seem to be supported.

Gitlab: closes #2206

¹ https://gnome.pages.gitlab.gnome.org/librsvg/Rsvg-2.0/migrating.html