core plugin xdot_style: consider epsilon <0.0005 when comparing pen widths
This branch is testing whether pen width is different from either the default or
what we cached last time. The pen width it then goes on to set in the output is
rounded to 3 decimal places. So consider anything that will round to an
identical value to be equal for the purposes of this comparison. This squashes a
-Wfloat-equal warning.