core plugin: remove 'color2str' and double buffering
The motivation for this change is removing a static buffer in `color2str` that
was impeding progress towards thread safety. Rather than tweak the calling
convention to avoid this, this change instead reworks how color serialization is
done here to remove the intermediate buffer altogether and write the color
directly into its target `agxbuf`.

Gitlab: #2558