core plugin fig_textspan: squash -Wbad-function-cast warning
GCC warns when you cast the result of a function to a different type that may
result in loss of precision, but does not warn if you split the operation into
an intermediate temporary like this. Perhaps `-Wbad-function-cast` should be
disabled in future as it seems unhelpful. In this case, if we have a string
length > 2⁵³ - 1 that can be stored in a `double`, something else has gone
deeply wrong.