add test case for loading Tclpathplan
The `xfail` for this test is left non-strict because it can pass when run
locally but fails in CI on some platforms. Future commits in this series will
narrow the exact failure conditions here. For reference, the current conditions
on its failure in CI are `is_cmake() or not (is_fedora() or is_rocky())`.