cgraph node_set_add: simplify set expansion logic
By choosing to expand the set at ≥ the watermark instead of > the watermark, we
can merge the two conditions here into one for a slight efficiency gain. The
previous separation was mainly driven by the goal of avoiding a division by 0
that was no longer a concern after the previous commit.